feat: refactor landing page structure and styles, consolidate components into main page
