+++ src/app/globals.css
@@ -5,7 +5,7 @@ html {
 }
 
 :root {
-  --tsk-primary-dark: #45084A;
+  --tsk-primary-dark: #45084a;
   --tsk-primary: #6f169e;
   --tsk-secondary: #a848c7;
   --tsk-light-1: #efd5f8;
@@ -103,61 +103,61 @@ h3 {
 }
 
 /*COLORS*/
-.bg-tsk-primary{
+.bg-tsk-primary {
   background-color: var(--tsk-primary);
 }
-.bg-tsk-primary-dark{
+.bg-tsk-primary-dark {
   background-color: var(--tsk-primary-dark);
 }
-.bg-tsk-secondary{
+.bg-tsk-secondary {
   background-color: var(--tsk-secondary);
 }
-.bg-tsk-light-1{
+.bg-tsk-light-1 {
   background-color: var(--tsk-light-1);
 }
-.bg-tsk-light-2{
+.bg-tsk-light-2 {
   background-color: var(--tsk-light-2);
 }
-.bg-foreground{
+.bg-foreground {
   background-color: var(--foreground);
 }
-.text-tsk-primary{
+.text-tsk-primary {
   color: var(--tsk-primary);
 }
-.text-tsk-primary-dark{
+.text-tsk-primary-dark {
   color: var(--tsk-primary-dark);
 }
-.text-tsk-secondary{
+.text-tsk-secondary {
   color: var(--tsk-secondary);
 }
-.text-tsk-light-1{
+.text-tsk-light-1 {
   color: var(--tsk-light-1);
 }
-.text-tsk-light-2{
+.text-tsk-light-2 {
   color: var(--tsk-light-2);
 }
-.text-foreground{
+.text-foreground {
   color: var(--foreground);
 }
-.border-tsk-primary{
+.border-tsk-primary {
   border-color: var(--tsk-primary-dark);
 }
-.border-tsk-primary-dark{
+.border-tsk-primary-dark {
   border-color: var(--tsk-primary-dark);
 }
-.border-tsk-secondary{
+.border-tsk-secondary {
   border-color: var(--tsk-secondary);
 }
 /*FONTS*/
-.font-heading{
+.font-heading {
   font-family: var(--font-heading);
 }
-.font-body{
+.font-body {
   font-family: var(--font-body);
 }
-.font-other{
+.font-other {
   font-family: var(--font-other);
 }
-.font-decorative{
+.font-decorative {
   font-family: var(--font-decorative);
 }
